Initial commit 🎉
This is a personal project. My goal is to learn more about os-development. The primary target of the operation system will be the raspberry pi 4B.
I use the rust-raspberrypi-OS-tutorials tutorial (https://github.com/rust-embedded/rust-raspberrypi-OS-tutorials) for reference, but will not strictly follow it, and also make use of knowledge from older projects / other references.